refactor: update message and pending flush models to use u64 for IDs

+++ src/database/entities/messages.rs
@@ -1,16 +1,15 @@
 use chrono::{DateTime, Utc};
 use sea_orm::entity::prelude::*;
-use serde::{Deserialize, Serialize};
 use serenity::all::*;
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
+#[derive(Clone, Debug, DeriveEntityModel)]
 #[sea_orm(table_name = "messages")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub message_id: i64,
-    pub user_id: i64,
-    pub guild_id: i64,
-    pub channel_id: i64,
+    pub message_id: u64,
+    pub user_id: u64,
+    pub guild_id: u64,
+    pub channel_id: u64,
     pub timestamp: DateTime<Utc>,
 }
 
@@ -21,19 +20,19 @@ impl ActiveModelBehavior for ActiveModel {}
 
 impl Model {
     pub fn message_id(&self) -> MessageId {
-        MessageId::from(self.message_id as u64)
+        MessageId::from(self.message_id)
     }
 
     pub fn user_id(&self) -> UserId {
-        UserId::from(self.user_id as u64)
+        UserId::from(self.user_id)
     }
 
     pub fn guild_id(&self) -> GuildId {
-        GuildId::from(self.guild_id as u64)
+        GuildId::from(self.guild_id)
     }
 
     pub fn channel_id(&self) -> ChannelId {
-        ChannelId::from(self.channel_id as u64)
+        ChannelId::from(self.channel_id)
     }
 
     pub fn timestamp(&self) -> DateTime<Utc> {
